Add script to redact raw emails
In some cases we may need to apply censor rules to raw emails stored on
disk. This script is an initial attempt to do this.

It might not be perfect and if you use this then you'll want to check
the redactions manually to ensure this is doing what you would expect.